Bumped dependency versions
* numpy from 1.23 to 2.1.0
* pyjls from 0.9.2 to 0.11.0
* pyjoulescope_driver from 1.5.3 to 1.7.0
* PySide 6.7.3 to 6.8.0

* Migrated release from Python 3.11 to 3.12

+++ joulescope_ui/json_plus.py
@@ -65,7 +65,7 @@ class CustomEncoder(json.JSONEncoder):
                          np.int8, np.int16, np.int32, np.int64,
                          np.uint8, np.uint16, np.uint32, np.uint64]:
             return int(obj)
-        elif type(obj) in [np.float_, np.float16, np.float32, np.float64]:
+        elif type(obj) in [np.float16, np.float32, np.float64, float]:
             return float(obj)
         else:
             _log.warning('Cannot serialize object: %s %s', type(obj), obj)
